Moved template parser tests into own module to reduce module size
