+++ src/template/parse.rs
@@ -1,3 +1,6 @@
+#[cfg(test)]
+mod tests;
+
 use color_eyre::eyre::{eyre, Result};
 use color_eyre::Report;
 
@@ -50,8 +53,7 @@ impl<'a> Parser<'a> {
 	}
 
 	fn next_top_level_block(&mut self) -> Option<Result<Block, DiagnositicBuilder>> {
-		// TODO-BM: cant handle these errors for now as information is missing
-		let Spanned { span, value: hint } = match self.next_block()? {
+		let Spanned { span, value: hint } = match self.blocks.next()? {
 			Ok(x) => x,
 			Err(err) => return Some(Err(err)),
 		};
@@ -68,6 +70,7 @@ impl<'a> Parser<'a> {
 			BlockHint::IfStart => self
 				.parse_if(span)
 				.map(|Spanned { span, value }| Block::new(span, BlockKind::If(value))),
+
 			// Illegal top level blocks
 			BlockHint::ElIf => Err(DiagnositicBuilder::new(DiagnositicLevel::Error)
 				.message("top-level `elif` block")
@@ -84,68 +87,6 @@ impl<'a> Parser<'a> {
 		};
 
 		Some(block)
-	}
-
-	fn try_recover(&mut self) -> bool {
-		todo!()
-	}
-
-	fn next_block(&mut self) -> Option<Result<Spanned<BlockHint>, DiagnositicBuilder>> {
-		let (span, hint, content) = match self.blocks.next()? {
-			Ok(x) => x,
-			Err(err) => {
-				return Some(Err(err));
-			}
-		};
-
-		if let Some(hint) = hint {
-			return Some(Ok(span.span(hint)));
-		}
-
-		// Check if its a text block (no opening and closing `{{\}}`)
-		if !matches!(content.as_bytes(), &[b'{', b'{', .., b'}', b'}']) {
-			return Some(Ok(span.span(BlockHint::Text)));
-		}
-
-		// Content without block opening and closing
-		let content = &content[2..content.len() - 2];
-
-		// Check for escaped
-		if let (Some(b'{'), Some(b'}')) = (content.as_bytes().get(0), content.as_bytes().last()) {
-			return Some(Ok(span.span(BlockHint::Escaped)));
-		}
-
-		// Check for comment
-		if let (Some(b"!--"), Some(b"--")) = (
-			content.as_bytes().get(..3),
-			content
-				.as_bytes()
-				.get(content.as_bytes().len().saturating_sub(3)..),
-		) {
-			return Some(Ok(span.span(BlockHint::Comment)));
-		}
-
-		// Check for if
-		if let Some(b"@if ") = content.as_bytes().get(..4) {
-			return Some(Ok(span.span(BlockHint::IfStart)));
-		}
-
-		// Check for elif
-		if let Some(b"@elif ") = content.as_bytes().get(..6) {
-			return Some(Ok(span.span(BlockHint::ElIf)));
-		}
-
-		// Check for else
-		if let Some(b"@else") = content.as_bytes().get(..5) {
-			return Some(Ok(span.span(BlockHint::Else)));
-		}
-
-		// Check for else
-		if let Some(b"@fi") = content.as_bytes().get(..3) {
-			return Some(Ok(span.span(BlockHint::IfEnd)));
-		}
-
-		Some(Ok(span.span(BlockHint::Variable)))
 	}
 
 	fn parse_text(&self, span: ByteSpan) -> Block {
@@ -204,7 +145,8 @@ impl<'a> Parser<'a> {
 			mut span,
 			value: mut hint,
 		} = self
-			.next_block()
+			.blocks
+			.next()
 			.ok_or_else(|| {
 				DiagnositicBuilder::new(DiagnositicLevel::Error)
 					.message("unexpected end of `if` block")
@@ -244,7 +186,8 @@ impl<'a> Parser<'a> {
 				span: _span,
 				value: _hint,
 			} = self
-				.next_block()
+				.blocks
+				.next()
 				.ok_or_else(|| {
 					DiagnositicBuilder::new(DiagnositicLevel::Error)
 						.message("unexpected end of `elif` block")
@@ -282,7 +225,8 @@ impl<'a> Parser<'a> {
 				span: _span,
 				value: _hint,
 			} = self
-				.next_block()
+				.blocks
+				.next()
 				.ok_or_else(|| {
 					DiagnositicBuilder::new(DiagnositicLevel::Error)
 						.message("unexpected end of `else` block")
@@ -427,9 +371,9 @@ impl<'a> Parser<'a> {
 	}
 
 	fn peek_block_hint(&self) -> Option<BlockHint> {
-		// TODO-BM: improve
-		let mut peek = self.clone();
-		peek.next_block()?.ok().map(|spanned| spanned.into_value())
+		// create a copy of the block iter to not mess up the state while peeking
+		let mut peek = self.blocks;
+		peek.next()?.ok().map(|spanned| spanned.into_value())
 	}
 }
 
@@ -614,7 +558,7 @@ impl<'a> BlockIter<'a> {
 }
 
 impl<'a> Iterator for BlockIter<'a> {
-	type Item = Result<(ByteSpan, Option<BlockHint>, &'a str), DiagnositicBuilder>;
+	type Item = Result<Spanned<BlockHint>, DiagnositicBuilder>;
 
 	fn next(&mut self) -> Option<Self::Item> {
 		let (mut span, hint) = match next_block(&self.content[self.index..])? {
@@ -642,561 +586,55 @@ impl<'a> Iterator for BlockIter<'a> {
 		span = span.offset(self.index as i32);
 		self.index = span.high().as_usize();
 
-		Some(Ok((span, hint, &self.content[span])))
-	}
-}
-
-#[cfg(test)]
-mod tests {
-	use pretty_assertions::assert_eq;
-
-	use super::*;
-	use crate::template::source::Source;
-	use crate::template::span::ByteSpan;
-
-	#[test]
-	fn parse_single_text() -> Result<()> {
-		let content = r#"Hello World this is a text block"#;
-
-		let source = Source::anonymous(content);
-		let mut parser = Parser::new(Session::new(source));
-		let block = parser
-			.next_top_level_block()
-			.expect("Found no block")
-			.expect("Encountered a parse error");
-
-		assert_eq!(
-			block,
-			Block::new(ByteSpan::new(0usize, content.len()), BlockKind::Text)
-		);
-
-		Ok(())
-	}
-
-	#[test]
-	fn parse_single_comment() -> Result<()> {
-		let content = r#"{{!-- Hello World this is a comment block --}}"#;
-
-		let source = Source::anonymous(content);
-		let mut parser = Parser::new(Session::new(source));
-		let block = parser
-			.next_top_level_block()
-			.expect("Found no block")
-			.expect("Encountered a parse error");
-
-		assert_eq!(
-			block,
-			Block::new(ByteSpan::new(0usize, content.len()), BlockKind::Comment)
-		);
-
-		Ok(())
-	}
-
-	#[test]
-	fn parse_single_escaped() -> Result<()> {
-		let content = r#"{{{ Hello World this is a comment block }}}"#;
-
-		let source = Source::anonymous(content);
-		let mut parser = Parser::new(Session::new(source));
-		let block = parser
-			.next_top_level_block()
-			.expect("Found no block")
-			.expect("Encountered a parse error");
-
-		assert_eq!(block.span(), &ByteSpan::new(0usize, content.len()));
-
-		let inner = ByteSpan::new(3usize, content.len() - 3);
-		assert_eq!(&content[inner], " Hello World this is a comment block ");
-		assert_eq!(block.kind(), &BlockKind::Escaped(inner));
-
-		Ok(())
-	}
-
-	#[test]
-	fn parse_single_var_default() -> Result<()> {
-		let content = r#"{{OS}}"#;
-
-		let source = Source::anonymous(content);
-		let mut parser = Parser::new(Session::new(source));
-		let block = parser
-			.next_top_level_block()
-			.expect("Found no block")
-			.expect("Encountered a parse error");
-
-		assert_eq!(block.span(), &ByteSpan::new(0usize, content.len()));
-
-		let name = ByteSpan::new(2usize, content.len() - 2);
-		assert_eq!(&content[name], "OS");
-		let envs = VarEnvSet([Some(VarEnv::Item), Some(VarEnv::Profile), None]);
-		assert_eq!(block.kind(), &BlockKind::Var(Var { envs, name }));
-
-		Ok(())
-	}
-
-	#[test]
-	fn parse_single_var_env() -> Result<()> {
-		let content = r#"{{$ENV}}"#;
-
-		let source = Source::anonymous(content);
-		let mut parser = Parser::new(Session::new(source));
-		let block = parser
-			.next_top_level_block()
-			.expect("Found no block")
-			.expect("Encountered a parse error");
-
-		assert_eq!(block.span(), &ByteSpan::new(0usize, content.len()));
-
-		let name = ByteSpan::new(3usize, content.len() - 2);
-		assert_eq!(&content[name], "ENV");
-		let envs = VarEnvSet([Some(VarEnv::Environment), None, None]);
-		assert_eq!(block.kind(), &BlockKind::Var(Var { envs, name }));
-
-		Ok(())
-	}
-
-	#[test]
-	fn parse_single_var_profile() -> Result<()> {
-		let content = r#"{{#PROFILE}}"#;
-
-		let source = Source::anonymous(content);
-		let mut parser = Parser::new(Session::new(source));
-		let block = parser
-			.next_top_level_block()
-			.expect("Found no block")
-			.expect("Encountered a parse error");
-
-		assert_eq!(block.span(), &ByteSpan::new(0usize, content.len()));
-
-		let name = ByteSpan::new(3usize, content.len() - 2);
-		assert_eq!(&content[name], "PROFILE");
-		let envs = VarEnvSet([Some(VarEnv::Profile), None, None]);
-		assert_eq!(block.kind(), &BlockKind::Var(Var { envs, name }));
-
-		Ok(())
-	}
-
-	#[test]
-	fn parse_single_var_item() -> Result<()> {
-		let content = r#"{{&ITEM}}"#;
-
-		let source = Source::anonymous(content);
-		let mut parser = Parser::new(Session::new(source));
-		let block = parser
-			.next_top_level_block()
-			.expect("Found no block")
-			.expect("Encountered a parse error");
-
-		assert_eq!(block.span(), &ByteSpan::new(0usize, content.len()));
-
-		let name = ByteSpan::new(3usize, content.len() - 2);
-		assert_eq!(&content[name], "ITEM");
-		let envs = VarEnvSet([Some(VarEnv::Item), None, None]);
-		assert_eq!(block.kind(), &BlockKind::Var(Var { envs, name }));
-
-		Ok(())
-	}
-
-	#[test]
-	fn parse_single_var_mixed() -> Result<()> {
-		let content = r#"{{$&#MIXED}}"#;
-
-		let source = Source::anonymous(content);
-		let mut parser = Parser::new(Session::new(source));
-		let block = parser
-			.next_top_level_block()
-			.expect("Found no block")
-			.expect("Encountered a parse error");
-
-		assert_eq!(block.span(), &ByteSpan::new(0usize, content.len()));
-
-		let name = ByteSpan::new(5usize, content.len() - 2);
-		assert_eq!(&content[name], "MIXED");
-		let envs = VarEnvSet([
-			Some(VarEnv::Environment),
-			Some(VarEnv::Item),
-			Some(VarEnv::Profile),
-		]);
-		assert_eq!(block.kind(), &BlockKind::Var(Var { envs, name }));
-
-		Ok(())
-	}
-
-	#[test]
-	fn parse_single_vars() -> Result<()> {
-		// duplicate variable environment
-		let content = r#"{{##OS}}"#;
-
-		let source = Source::anonymous(content);
-		let mut parser = Parser::new(Session::new(source));
-		let block = parser
-			.next_top_level_block()
-			.ok_or(eyre!("No block found"))?;
-
-		assert!(block.is_err());
-
-		Ok(())
-	}
-
-	#[test]
-	fn parse_single_if_eq() -> Result<()> {
-		let content = r#"{{@if {{OS}} == "windows"}}{{@fi}}"#;
-
-		let source = Source::anonymous(content);
-		let mut parser = Parser::new(Session::new(source));
-		let block = parser
-			.next_top_level_block()
-			.expect("Found no block")
-			.expect("Encountered a parse error");
-
-		assert_eq!(block.span(), &ByteSpan::new(0usize, content.len()));
-
-		let if_span = ByteSpan::new(0usize, 27usize);
-		assert_eq!(&content[if_span], r#"{{@if {{OS}} == "windows"}}"#);
-
-		let name = ByteSpan::new(8usize, 10usize);
-		assert_eq!(&content[name], "OS");
-		let envs = VarEnvSet([Some(VarEnv::Item), Some(VarEnv::Profile), None]);
-
-		let op = IfOp::Eq;
-
-		let other = ByteSpan::new(17usize, 24usize);
-		assert_eq!(&content[other], "windows");
-
-		let end_span = ByteSpan::new(27usize, 34usize);
-		assert_eq!(&content[end_span], r#"{{@fi}}"#);
-
-		assert_eq!(
-			block.kind(),
-			&BlockKind::If(If {
-				head: (
-					if_span.span(IfExpr::Compare {
-						var: Var { envs, name },
-						op,
-						other
-					}),
-					vec![]
-				),
-				elifs: vec![],
-				els: None,
-				end: end_span
-			})
-		);
-
-		Ok(())
-	}
-
-	#[test]
-	fn parse_single_if_neq() -> Result<()> {
-		let content = r#"{{@if {{OS}} != "windows"}}{{@fi}}"#;
-
-		let source = Source::anonymous(content);
-		let mut parser = Parser::new(Session::new(source));
-		let block = parser
-			.next_top_level_block()
-			.expect("Found no block")
-			.expect("Encountered a parse error");
-
-		assert_eq!(block.span(), &ByteSpan::new(0usize, content.len()));
-
-		let if_span = ByteSpan::new(0usize, 27usize);
-		assert_eq!(&content[if_span], r#"{{@if {{OS}} != "windows"}}"#);
-
-		let name = ByteSpan::new(8usize, 10usize);
-		assert_eq!(&content[name], "OS");
-		let envs = VarEnvSet([Some(VarEnv::Item), Some(VarEnv::Profile), None]);
-
-		let op = IfOp::NotEq;
-
-		let other = ByteSpan::new(17usize, 24usize);
-		assert_eq!(&content[other], "windows");
-
-		let end_span = ByteSpan::new(27usize, 34usize);
-		assert_eq!(&content[end_span], r#"{{@fi}}"#);
-
-		assert_eq!(
-			block.kind(),
-			&BlockKind::If(If {
-				head: (
-					if_span.span(IfExpr::Compare {
-						var: Var { envs, name },
-						op,
-						other
-					}),
-					vec![]
-				),
-				elifs: vec![],
-				els: None,
-				end: end_span
-			})
-		);
-
-		Ok(())
-	}
-
-	#[test]
-	fn parse_single_if_exists() -> Result<()> {
-		let content = r#"{{@if {{$#EXISTS}}}}{{@fi}}"#;
-
-		let source = Source::anonymous(content);
-		let mut parser = Parser::new(Session::new(source));
-		let block = parser
-			.next_top_level_block()
-			.expect("Found no block")
-			.expect("Encountered a parse error");
-
-		assert_eq!(block.span(), &ByteSpan::new(0usize, content.len()));
-
-		let if_span = ByteSpan::new(0usize, 20usize);
-		assert_eq!(&content[if_span], r#"{{@if {{$#EXISTS}}}}"#);
-
-		let name = ByteSpan::new(10usize, 16usize);
-		assert_eq!(&content[name], "EXISTS");
-		let envs = VarEnvSet([Some(VarEnv::Environment), Some(VarEnv::Profile), None]);
-
-		let end_span = ByteSpan::new(20usize, 27usize);
-		assert_eq!(&content[end_span], r#"{{@fi}}"#);
-
-		assert_eq!(
-			block.kind(),
-			&BlockKind::If(If {
-				head: (
-					if_span.span(IfExpr::Exists {
-						var: Var { envs, name }
-					}),
-					vec![]
-				),
-				elifs: vec![],
-				els: None,
-				end: end_span
-			})
-		);
-
-		Ok(())
-	}
-
-	#[test]
-	fn find_blocks() {
-		let content = r#"{{ Hello World }} {{{ Escaped {{ }} }} }}}
-		{{!-- Hello World {{}} {{{ asdf }}} this is a comment --}}
-		{{@if {{}} }} }}
-		"#;
-
-		println!("{}", content);
-
-		let iter = BlockIter::new(content);
-
-		// Hello World
-		// Text: SPACE
-		// Escaped
-		// Text: LF SPACES
-		// Comment
-		// Text: LF SPACES
-		// If
-		// Text: Closing LF SPACES
-		assert_eq!(iter.count(), 8);
-	}
-
-	#[test]
-	fn find_blocks_unicode() {
-		let content = "\u{1f600}{{{ \u{1f600} }}}\u{1f600}";
-
-		let iter = BlockIter::new(content);
-
-		// Text: Smiley
-		// Escaped
-		// Text: Smiley
-		assert_eq!(iter.count(), 3);
-	}
-
-	#[test]
-	fn parse_comment() -> Result<()> {
-		let content = r#"{{!-- Hello World this {{}} is a comment {{{{{{ }}}--}}"#;
-
-		let source = Source::anonymous(content);
-		let mut parser = Parser::new(Session::new(source));
-		let token = parser
-			.next_top_level_block()
-			.expect("Found no block")
-			.expect("Encountered a parse error");
-
-		assert_eq!(
-			token,
-			Block::new(ByteSpan::new(0usize, content.len()), BlockKind::Comment)
-		);
-
-		Ok(())
-	}
-
-	#[test]
-	fn parse_escaped() -> Result<()> {
-		let content = r#"{{{!-- Hello World this {{}} is a comment {{{{{{ }}--}}}"#;
-
-		let source = Source::anonymous(content);
-		let mut parser = Parser::new(Session::new(source));
-		let token = parser
-			.next_top_level_block()
-			.expect("Found no block")
-			.expect("Encountered a parse error");
-
-		assert_eq!(
-			token,
-			Block::new(
-				ByteSpan::new(0usize, content.len()),
-				BlockKind::Escaped(ByteSpan::new(3usize, content.len() - 3))
-			)
-		);
-
-		Ok(())
-	}
-
-	#[test]
-	fn parse_if_cmp() -> Result<()> {
-		let content = r#"{{@if {{&OS}} == "windows" }}
-		DEMO
-		{{@elif {{&OS}} == "linux"  }}
-		LINUX
-		{{@else}}
-		ASD
-		{{@fi}}"#;
-
-		let source = Source::anonymous(content);
-		let mut parser = Parser::new(Session::new(source));
-		let token = parser
-			.next_top_level_block()
-			.expect("Found no block")
-			.expect("Encountered a parse error");
-
-		assert_eq!(token.span, ByteSpan::new(0usize, content.len()));
-		println!("{:#?}", &token.kind);
-
-		Ok(())
-	}
-
-	#[test]
-	fn parse_if_cmp_nested() -> Result<()> {
-		let content = r#"{{@if {{&OS}} == "windows" }}
-		{{!-- This is a nested comment --}}
-		{{{ Escaped {{}} }}}
-		{{@elif {{&OS}} == "linux"  }}
-		{{!-- Below is a nested variable --}}
-		{{ OS }}
-		{{@else}}
-		ASD
-		{{@fi}}"#;
-
-		let source = Source::anonymous(content);
-		let mut parser = Parser::new(Session::new(source));
-		let token = parser
-			.next_top_level_block()
-			.expect("Found no block")
-			.expect("Encountered a parse error");
-
-		assert_eq!(token.span, ByteSpan::new(0usize, content.len()));
-		println!("{:#?}", &token.kind);
-
-		Ok(())
-	}
-
-	#[test]
-	fn parse_if_exists() -> Result<()> {
-		let content = r#"{{@if {{&OS}}  }}
-		DEMO
-		ASD
-		{{@fi}}"#;
-
-		let source = Source::anonymous(content);
-		let mut parser = Parser::new(Session::new(source));
-		let token = parser
-			.next_top_level_block()
-			.expect("Found no block")
-			.expect("Encountered a parse error");
-
-		assert_eq!(token.span, ByteSpan::new(0usize, content.len()));
-		println!("{:#?}", &token.kind);
-
-		Ok(())
-	}
-
-	#[test]
-	fn parse_if_mixed() -> Result<()> {
-		let content = r#"{{@if {{OS}}}}
-	print("No value for variable `OS` set")
-{{@elif {{&OS}} != "windows"}}
-	print("OS is not windows")
-{{@elif {{OS}} == "windows"}}
-	{{{!-- This is a nested comment. Below it is a nested variable block. --}}}
-	print("OS is {{OS}}")
-{{@else}}
-	{{{!-- This is a nested comment. --}}}
-	print("Can never get here. {{{ {{OS}} is neither `windows` nor not `windows`. }}}")
-{{@fi}}"#;
-
-		let source = Source::anonymous(content);
-		let mut parser = Parser::new(Session::new(source));
-		let token = parser
-			.next_top_level_block()
-			.expect("Found no block")
-			.expect("Encountered a parse error");
-
-		assert_eq!(token.span, ByteSpan::new(0usize, content.len()));
-		println!("{:#?}", &token.kind);
-
-		Ok(())
-	}
-
-	#[test]
-	fn parse_variables() -> Result<()> {
-		assert_eq!(
-			parse_var("$#&FOO_BAR", 0)?,
-			Var {
-				envs: VarEnvSet([
-					Some(VarEnv::Environment),
-					Some(VarEnv::Profile),
-					Some(VarEnv::Item)
-				]),
-				name: ByteSpan::new(3usize, 10usize),
-			}
-		);
-
-		assert_eq!(
-			parse_var("&BAZ_1", 0)?,
-			Var {
-				envs: VarEnvSet([Some(VarEnv::Item), None, None]),
-				name: ByteSpan::new(1usize, 6usize),
-			}
-		);
-
-		assert_eq!(
-			parse_var("$#&FOO_BAR", 10)?,
-			Var {
-				envs: VarEnvSet([
-					Some(VarEnv::Environment),
-					Some(VarEnv::Profile),
-					Some(VarEnv::Item)
-				]),
-				name: ByteSpan::new(13usize, 20usize),
-			}
-		);
-
-		// invalid env / var_name
-		assert!(parse_var("!FOO_BAR", 10).is_err());
-		// duplicate env
-		assert!(parse_var("&&FOO_BAR", 0).is_err());
-
-		Ok(())
-	}
-
-	#[test]
-	fn parse_others() -> Result<()> {
-		assert_eq!(parse_other("\"BAZ_1\"", 0)?, ByteSpan::new(1usize, 6usize));
-		assert_eq!(
-			parse_other("This is a test \"Hello World How are you today\"", 0)?,
-			ByteSpan::new(16usize, 45usize)
-		);
-
-		assert!(parse_other("This is a test \"Hello World How are you today", 0).is_err());
-		assert!(parse_other("This is a test", 0).is_err());
-
-		Ok(())
+		let content = &self.content[span];
+
+		if let Some(hint) = hint {
+			return Some(Ok(span.span(hint)));
+		}
+
+		// Check if its a text block (no opening and closing `{{\}}`)
+		if !matches!(content.as_bytes(), &[b'{', b'{', .., b'}', b'}']) {
+			return Some(Ok(span.span(BlockHint::Text)));
+		}
+
+		// Content without block opening and closing
+		let content = &content[2..content.len() - 2];
+
+		// Check for escaped
+		if let (Some(b'{'), Some(b'}')) = (content.as_bytes().get(0), content.as_bytes().last()) {
+			return Some(Ok(span.span(BlockHint::Escaped)));
+		}
+
+		// Check for comment
+		if let (Some(b"!--"), Some(b"--")) = (
+			content.as_bytes().get(..3),
+			content
+				.as_bytes()
+				.get(content.as_bytes().len().saturating_sub(3)..),
+		) {
+			return Some(Ok(span.span(BlockHint::Comment)));
+		}
+
+		// Check for if
+		if let Some(b"@if ") = content.as_bytes().get(..4) {
+			return Some(Ok(span.span(BlockHint::IfStart)));
+		}
+
+		// Check for elif
+		if let Some(b"@elif ") = content.as_bytes().get(..6) {
+			return Some(Ok(span.span(BlockHint::ElIf)));
+		}
+
+		// Check for else
+		if let Some(b"@else") = content.as_bytes().get(..5) {
+			return Some(Ok(span.span(BlockHint::Else)));
+		}
+
+		// Check for else
+		if let Some(b"@fi") = content.as_bytes().get(..3) {
+			return Some(Ok(span.span(BlockHint::IfEnd)));
+		}
+
+		Some(Ok(span.span(BlockHint::Variable)))
 	}
 }
